authorization

+++ app/Http/Controllers/ListingController.php
@@ -62,6 +62,11 @@ class ListingController extends Controller
 
     // Update listing data
     public function update(Request $request, Listing $listing){ 
+
+        if($listing->user_id != auth()->id()){
+            abort('403', 'Unauthorized action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -84,13 +89,16 @@ class ListingController extends Controller
     
     // Delete listing
     public function destroy(Listing $listing) {
+
+        if($listing->user_id != auth()->id()){
+            abort('403', 'Unauthorized action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfuly!');
     }
     public function manage(){
         return view('listings.manage', ['listings' => auth()->user()->listings]);
     }
-
-    
-    
+   
 }
